+++ public/prod.config.template.js
@@ -1,0 +1,17 @@
+/*
+* public/prod.config.js
+*
+* Production configuration.
+*
+* TEMPLATE: edit to your liking! You can also omit the contents for no additional production tooling.
+*/
+
+// Airbrake
+const ops = {
+  projectId: undefined,
+  projectKey: undefined
+}
+
+export {
+  ops
+}