+++ rollup.config.prod.js
@@ -40,6 +40,9 @@ import { prodIndexPlugin } from './tools/prod-index-filter'
 //import { version } from './package.json'
 const { version } = require('./package.json');    // this works, 'import' didn't
 
+import dotenv from 'dotenv'
+dotenv.config();    // apply '.env.production' file (if there) -> 'process.env'
+
 /*
 * Note: The order of the plugins does sometimes matter.
 */